requirements.txt
requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=0.20.0
PyYAML>=6.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
rich>=12.0.0
aiohttp>=3.8.0
asyncio-contextmanager>=1.0.0